Language.SQL.SimpleSQL.Dialect

Documentation

data Dialect

Used to set the dialect used for parsing and pretty printing, very unfinished at the moment.

The keyword handling works as follows:

There is a list of reserved keywords. These will never parse as anything other than as a keyword, unless they are in one of the other lists.

There is a list of 'identifier' keywords. These are reserved keywords, with an exception that they will parse as an identifier in a scalar expression. They won't parse as identifiers in other places, e.g. column names or aliases.

There is a list of 'app' keywords. These are reserved keywords, with an exception that they will also parse in an 'app-like' construct - a regular function call, or any of the aggregate and window variations.

There is a list of special type names. This list serves two purposes - it is a list of the reserved keywords which are also type names, and it is a list of all the multi word type names.

Every keyword should appear in the keywords lists, and then you can add them to the other lists if you want exceptions. Most things that refer to functions, types or variables that are keywords in the ansi standard, can be removed from the keywords lists completely with little effect. With most of the actual SQL keywords, removing them from the keyword list will result in lots of valid syntax no longer parsing (and probably bad parse error messages too).

In the code, all special syntax which looks identical to regular identifiers or function calls (apart from the name), is treated like a regular identifier or function call.

Constructors

Dialect

Fields

diKeywords :: [String]

reserved keywords

diIdentifierKeywords :: [String]

keywords with identifier exception

diAppKeywords :: [String]

keywords with app exception

diSpecialTypeNames :: [String]

keywords with type exception plus all the type names which are multiple words

diFetchFirst :: Bool

allow ansi fetch first syntax

diLimit :: Bool

allow limit keyword (mysql, postgres, ...)

diOdbc :: Bool

allow parsing ODBC syntax

diBackquotedIden :: Bool

allow quoting identifiers with `backquotes`

diSquareBracketQuotedIden :: Bool

allow quoting identifiers with [square brackets]

diAtIdentifier :: Bool

allow identifiers with a leading at @example

diHashIdentifier :: Bool

allow identifiers with a leading # #example

diPositionalArg :: Bool

allow positional identifiers like this: $1

diDollarString :: Bool

allow postgres style dollar strings

diEString :: Bool

allow strings with an e - e"example"

diPostgresSymbols :: Bool

allow postgres style symbols

diSqlServerSymbols :: Bool

allow sql server style symbols

ansi2011 :: Dialect

ansi sql 2011 dialect

mysql :: Dialect

mysql dialect

postgres :: Dialect

postgresql dialect

oracle :: Dialect

oracle dialect

sqlserver :: Dialect

microsoft sql server dialect