Language.SQL.SimpleSQL.Lexer

Description

This is the module contains a Lexer for SQL.

Documentation

data Token

Represents a lexed token

Constructors

Symbol String
A symbol is one of the following

multi char symbols <> == != ||

single char symbols: * + - ^ / % ~ & | ? ( ) [ ] , ; ( )

Identifier String
This is an identifier or keyword.

QIdentifier String
This is an identifier quoted with "

UQIdentifier String
This is an identifier quoted with u&"

DQIdentifier String String String
This is a dialect specific quoted identifier with the quote characters explicit. The first and second fields are the starting and ending quote characters.

HostParam String
This is a host param symbol, e.g. :param

SqlString String
This is a string literal.

CSSqlString String String
This is a character set string literal. The first field is the character set (one of nNbBxX, or u&, U&).

SqlNumber String
A number literal (integral or otherwise), stored in original format unchanged

Whitespace String
Whitespace, one or more of space, tab or newline.

LineComment String
A commented line using --, contains every character starting with the '--' and including the terminating newline character if there is one - this will be missing if the last line in the source is a line comment with no trailing newline

BlockComment String
A block comment, /* stuff */, includes the comment delimiters

lexSQL

Arguments

:: Dialect
dialect of SQL to use

-> FilePath
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> String
the SQL source to lex

-> Either ParseError [((String, Int, Int), Token)]

Lex some SQL to a list of tokens.

prettyToken :: Dialect -> Token -> String

Pretty printing, if you lex a bunch of tokens, then pretty print them, should should get back exactly the same string

prettyTokens :: Dialect -> [Token] -> String

data ParseError

Type to represent parse errors.

Constructors

ParseError

Fields

peErrorString :: String

contains the error message

peFilename :: FilePath

filename location for the error

pePosition :: (Int, Int)

line number and column number location for the error

peFormattedError :: String

formatted error with the position, error message and source context

data Dialect

Used to set the dialect used for parsing and pretty printing, very unfinished at the moment.

Constructors

SQL2011

MySQL