Language.SQL.SimpleSQL.Parse

Description

This is the module with the parser functions.

Documentation

parseQueryExpr

Arguments

:: Dialect
dialect of SQL to use

-> Text
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> Text
the SQL source to parse

-> Either ParseError QueryExpr

Parses a query expr, trailing semicolon optional.

parseScalarExpr

Arguments

:: Dialect
dialect of SQL to use

-> Text
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> Text
the SQL source to parse

-> Either ParseError ScalarExpr

Parses a scalar expression.

parseStatement

Arguments

:: Dialect
dialect of SQL to use

-> Text
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> Text
the SQL source to parse

-> Either ParseError Statement

Parses a statement, trailing semicolon optional.

parseStatements

Arguments

:: Dialect
dialect of SQL to use

-> Text
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> Text
the SQL source to parse

-> Either ParseError [Statement]

Parses a list of statements, with semi colons between them. The final semicolon is optional.

data ParseError

Constructors

LexError ParseError

ParseError (ParseErrorBundle SQLStream Void)

prettyError :: ParseError -> Text

ansi2011 :: Dialect

ansi sql 2011 dialect