Language.SQL.SimpleSQL.Pretty

Description

These is the pretty printing functions, which produce SQL source from ASTs. The code attempts to format the output in a readable way.

Documentation

prettyQueryExpr :: QueryExpr -> String

Convert a query expr ast to concrete syntax.

prettyScalarExpr :: ScalarExpr -> String

Convert a scalar expr ast to concrete syntax.

prettyQueryExprs :: [QueryExpr] -> String

Convert a list of query exprs to concrete syntax. A semi colon is inserted after each query expr.