Language.SQL.SimpleSQL.Parser

Description

This is the module with the parser functions.

Documentation

parseQueryExpr

Arguments

:: Dialect
dialect of SQL to use

-> FilePath
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> String
the SQL source to parse

-> Either ParseError QueryExpr

Parses a query expr, trailing semicolon optional.

parseValueExpr

Arguments

:: Dialect
dialect of SQL to use

-> FilePath
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> String
the SQL source to parse

-> Either ParseError ValueExpr

Parses a value expression.

parseStatement

Arguments

:: Dialect
dialect of SQL to use

-> FilePath
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> String
the SQL source to parse

-> Either ParseError Statement

Parses a statement, trailing semicolon optional.

parseStatements

Arguments

:: Dialect
dialect of SQL to use

-> FilePath
filename to use in error messages

-> Maybe (Int, Int)
line number and column number of the first character in the source to use in error messages

-> String
the SQL source to parse

-> Either ParseError [Statement]

Parses a list of statements, with semi colons between them. The final semicolon is optional.

data ParseError

Type to represent parse errors.

Constructors

ParseError

Fields

peErrorString :: String

contains the error message

peFilename :: FilePath

filename location for the error

pePosition :: (Int, Int)

line number and column number location for the error

peFormattedError :: String

formatted error with the position, error message and source context