Language.SQL.SimpleSQL.Pretty

Description

These is the pretty printing functions, which produce SQL source from ASTs. The code attempts to format the output in a readable way.

Documentation

prettyQueryExpr :: Dialect -> QueryExpr -> String

Convert a query expr ast to concrete syntax.

prettyScalarExpr :: Dialect -> ScalarExpr -> String

Convert a value expr ast to concrete syntax.

prettyStatement :: Dialect -> Statement -> String

Convert a statement ast to concrete syntax.

prettyStatements :: Dialect -> [Statement] -> String

Convert a list of statements to concrete syntax. A semicolon is inserted after each statement.