Language.SQL.SimpleSQL.Syntax

Description

The AST for SQL.

Scalar expressions

data ScalarExpr

Represents a value expression. This is used for the expressions in select lists. It is also used for expressions in where, group by, having, order by and so on.

Constructors

NumLit String
a numeric literal optional decimal point, e+- integral exponent, e.g

10

10.

.1

10.1

1e5

12.34e-6

StringLit String String String
string literal, with the start and end quote e.g. test -> StringLit "'" "'" "test"

IntervalLit
text of interval literal, units of interval precision, e.g. interval 3 days (3)

Fields

ilSign :: Maybe Sign

if + or - used

ilLiteral :: String

literal text

ilFrom :: IntervalTypeField

ilTo :: Maybe IntervalTypeField

TypedLit TypeName String
prefix 'typed literal', e.g. int '42'

Iden [Name]
identifier with parts separated by dots

Star
star, as in select *, t.*, count(*)

Parameter
Represents a ? in a parameterized query

PositionalArg Int
Represents an e.g. $1 in a parameterized query

HostParameter String (Maybe String)
represents a host parameter, e.g. :a. The Maybe String is for the indicator, e.g. :var indicator :nl

BinOp ScalarExpr [Name] ScalarExpr
Infix binary operators. This is used for symbol operators (a + b), keyword operators (a and b) and multiple keyword operators (a is similar to b)

PrefixOp [Name] ScalarExpr
Prefix unary operators. This is used for symbol operators, keyword operators and multiple keyword operators.

PostfixOp [Name] ScalarExpr
Postfix unary operators. This is used for symbol operators, keyword operators and multiple keyword operators.

SpecialOp [Name] [ScalarExpr]
Used for ternary, mixfix and other non orthodox operators. Currently used for row constructors, and for between.

App [Name] [ScalarExpr]
function application (anything that looks like c style function application syntactically)

AggregateApp
aggregate application, which adds distinct or all, and order by, to regular function application

Fields

aggName :: [Name]

aggregate function name

aggDistinct :: SetQuantifier

distinct

aggArgs :: [ScalarExpr]

args

aggOrderBy :: [SortSpec]

order by

aggFilter :: Maybe ScalarExpr

filter

AggregateAppGroup
aggregates with within group

Fields

aggName :: [Name]

aggregate function name

aggArgs :: [ScalarExpr]

args

aggGroup :: [SortSpec]

within group

WindowApp
window application, which adds over (partition by a order by b) to regular function application. Explicit frames are not currently supported

Fields

wnName :: [Name]

window function name

wnArgs :: [ScalarExpr]

args

wnPartition :: [ScalarExpr]

partition by

wnOrderBy :: [SortSpec]

order by

wnFrame :: Maybe Frame

frame clause

SpecialOpK [Name] (Maybe ScalarExpr) [(String, ScalarExpr)]
Used for the operators which look like functions except the arguments are separated by keywords instead of commas. The maybe is for the first unnamed argument if it is present, and the list is for the keyword argument pairs.

Cast ScalarExpr TypeName
cast(a as typename)

Case
case expression. both flavours supported

Fields

caseTest :: Maybe ScalarExpr

test value

caseWhens :: [([ScalarExpr], ScalarExpr)]

when branches

caseElse :: Maybe ScalarExpr

else value

Parens ScalarExpr

In Bool ScalarExpr InPredValue
in list literal and in subquery, if the bool is false it means not in was used ('a not in (1,2)')

SubQueryExpr SubQueryExprType QueryExpr
exists, all, any, some subqueries

QuantifiedComparison ScalarExpr [Name] CompPredQuantifier QueryExpr

Match ScalarExpr Bool QueryExpr

Array ScalarExpr [ScalarExpr]
represents an array access expression, or an array ctor e.g. a[3]. The first scalarExpr is the array, the second is the subscripts/ctor args

ArrayCtor QueryExpr
this is used for the query expression version of array constructors, e.g. array(select * from t)

Collate ScalarExpr [Name]

MultisetBinOp ScalarExpr SetOperatorName SetQuantifier ScalarExpr

MultisetCtor [ScalarExpr]

MultisetQueryCtor QueryExpr

NextValueFor [Name]

VEComment [Comment] ScalarExpr

OdbcLiteral OdbcLiteralType String
an odbc literal e.g. {d '2000-01-01'}

OdbcFunc ScalarExpr
an odbc function call e.g. {fn CHARACTER_LENGTH(test)}

data Name

Represents an identifier name, which can be quoted or unquoted. examples:

test -> Name Nothing "test"

"test" -> Name (Just """,""") "test"

something -> Name (Just ("`","`") "something"

[ms] -> Name (Just ("[","]") "ms"

Constructors

Name (Maybe (String, String)) String

data TypeName

Represents a type name, used in casts.

Constructors

TypeName [Name]

PrecTypeName [Name] Integer

PrecScaleTypeName [Name] Integer Integer

PrecLengthTypeName [Name] Integer (Maybe PrecMultiplier) (Maybe PrecUnits)

CharTypeName [Name] (Maybe Integer) [Name] [Name]

TimeTypeName [Name] (Maybe Integer) Bool

RowTypeName [(Name, TypeName)]

IntervalTypeName IntervalTypeField (Maybe IntervalTypeField)

ArrayTypeName TypeName (Maybe Integer)

MultisetTypeName TypeName

data IntervalTypeField

Constructors

Itf String (Maybe (Integer, Maybe Integer))

data Sign

Constructors

Plus

Minus

data PrecMultiplier

Constructors

PrecK

PrecM

PrecG

PrecT

PrecP

data PrecUnits

Constructors

PrecCharacters

PrecOctets

data SetQuantifier

Represents the Distinct or All keywords, which can be used before a select list, in an aggregate/window function application, or in a query expression set operator.

Constructors

SQDefault

Distinct

All

data SortSpec

Represents one field in an order by list.

Constructors

SortSpec ScalarExpr Direction NullsOrder

data Direction

The direction for a column in order by.

Constructors

DirDefault

Asc

Desc

data NullsOrder

Represents 'nulls first' or 'nulls last' in an order by clause.

Constructors

NullsOrderDefault

NullsFirst

NullsLast

data InPredValue

Used for 'expr in (scalar expression list)', and 'expr in (subquery)' syntax.

Constructors

InList [ScalarExpr]

InQueryExpr QueryExpr

data SubQueryExprType

A subquery in a scalar expression.

Constructors

SqExists
exists (query expr)

SqUnique
unique (query expr)

SqSq
a scalar subquery

data CompPredQuantifier

Constructors

CPAny

CPSome

CPAll

data Frame

Represents the frame clause of a window this can be [range | rows] frame_start or [range | rows] between frame_start and frame_end

Constructors

FrameFrom FrameRows FramePos

FrameBetween FrameRows FramePos FramePos

data FrameRows

Represents whether a window frame clause is over rows or ranges.

Constructors

FrameRows

FrameRange

data FramePos

represents the start or end of a frame

Constructors

UnboundedPreceding

Preceding ScalarExpr

Current

Following ScalarExpr

UnboundedFollowing

data OdbcLiteralType

the type of an odbc literal (e.g. {d '2000-01-01'}), correpsonding to the letter after the opening {

Constructors

OLDate

OLTime

OLTimestamp

Query expressions

data QueryExpr

Represents a query expression, which can be:

a regular select;

a set operator (union, except, intersect);

a common table expression (with);

a table value constructor (values (1,2),(3,4)); or

an explicit table (table t).

Constructors

Select

Fields

qeSetQuantifier :: SetQuantifier

qeSelectList :: [(ScalarExpr, Maybe Name)]

the expressions and the column aliases

qeFrom :: [TableRef]

qeWhere :: Maybe ScalarExpr

qeGroupBy :: [GroupingExpr]

qeHaving :: Maybe ScalarExpr

qeOrderBy :: [SortSpec]

qeOffset :: Maybe ScalarExpr

qeFetchFirst :: Maybe ScalarExpr

QueryExprSetOp

Fields

qe0 :: QueryExpr

qeCombOp :: SetOperatorName

qeSetQuantifier :: SetQuantifier

qeCorresponding :: Corresponding

qe1 :: QueryExpr

With

Fields

qeWithRecursive :: Bool

qeViews :: [(Alias, QueryExpr)]

qeQueryExpression :: QueryExpr

Values [[ScalarExpr]]

Table [Name]

QEComment [Comment] QueryExpr

makeSelect :: QueryExpr

Helper/'default' value for query exprs to make creating query expr values a little easier. It is defined like this:

makeSelect :: QueryExpr
makeSelect = Select {qeSetQuantifier = SQDefault
                    ,qeSelectList = []
                    ,qeFrom = []
                    ,qeWhere = Nothing
                    ,qeGroupBy = []
                    ,qeHaving = Nothing
                    ,qeOrderBy = []
                    ,qeOffset = Nothing
                    ,qeFetchFirst = Nothing}

data SetOperatorName

Query expression set operators.

Constructors

Union

Except

Intersect

data Corresponding

Corresponding, an option for the set operators.

Constructors

Corresponding

Respectively

data Alias

Represents an alias for a table valued expression, used in with queries and in from alias, e.g. select a from t u, select a from t u(b), with a(c) as select 1, select * from a.

Constructors

Alias Name (Maybe [Name])

data GroupingExpr

Represents an item in a group by clause.

Constructors

GroupingParens [GroupingExpr]

Cube [GroupingExpr]

Rollup [GroupingExpr]

GroupingSets [GroupingExpr]

SimpleGroup ScalarExpr

From

data TableRef

Represents a entry in the csv of tables in the from clause.

Constructors

TRSimple [Name]
from t / from s.t

TRJoin TableRef Bool JoinType TableRef (Maybe JoinCondition)
from a join b, the bool is true if natural was used

TRParens TableRef
from (a)

TRAlias TableRef Alias
from a as b(c,d)

TRQueryExpr QueryExpr
from (query expr)

TRFunction [Name] [ScalarExpr]
from function(args)

TRLateral TableRef
from lateral t

TROdbc TableRef
ODBC {oj t1 left outer join t2 on expr} syntax

data JoinType

The type of a join.

Constructors

JInner

JLeft

JRight

JFull

JCross

data JoinCondition

The join condition.

Constructors

JoinOn ScalarExpr
on expr

JoinUsing [Name]
using (column list)

Statements

data Statement

Constructors

CreateSchema [Name]

DropSchema [Name] DropBehaviour

CreateTable [Name] [TableElement]

AlterTable [Name] AlterTableAction

DropTable [Name] DropBehaviour

CreateView Bool [Name] (Maybe [Name]) QueryExpr (Maybe CheckOption)

DropView [Name] DropBehaviour

CreateDomain [Name] TypeName (Maybe ScalarExpr) [(Maybe [Name], ScalarExpr)]

AlterDomain [Name] AlterDomainAction

DropDomain [Name] DropBehaviour

CreateAssertion [Name] ScalarExpr

DropAssertion [Name] DropBehaviour

CreateSequence [Name] [SequenceGeneratorOption]

AlterSequence [Name] [SequenceGeneratorOption]

DropSequence [Name] DropBehaviour

SelectStatement QueryExpr

Delete [Name] (Maybe Name) (Maybe ScalarExpr)

Truncate [Name] IdentityRestart

Insert [Name] (Maybe [Name]) InsertSource

Update [Name] (Maybe Name) [SetClause] (Maybe ScalarExpr)

GrantPrivilege [PrivilegeAction] PrivilegeObject [Name] GrantOption

GrantRole [Name] [Name] AdminOption

CreateRole Name

DropRole Name

RevokePrivilege GrantOptionFor [PrivilegeAction] PrivilegeObject [Name] DropBehaviour

RevokeRole AdminOptionFor [Name] [Name] DropBehaviour

StartTransaction

Savepoint Name

ReleaseSavepoint Name

Commit

Rollback (Maybe Name)

StatementComment [Comment]

data DropBehaviour

Constructors

Restrict

Cascade

DefaultDropBehaviour

data IdentityRestart

Constructors

ContinueIdentity

RestartIdentity

DefaultIdentityRestart

data InsertSource

Constructors

InsertQuery QueryExpr

DefaultInsertValues

data SetClause

Constructors

Set [Name] ScalarExpr

SetMultiple [[Name]] [ScalarExpr]

data TableElement

Constructors

TableColumnDef ColumnDef

TableConstraintDef (Maybe [Name]) TableConstraint

data ColumnDef

Constructors

ColumnDef Name TypeName (Maybe DefaultClause) [ColConstraintDef]

data DefaultClause

Constructors

DefaultClause ScalarExpr

IdentityColumnSpec IdentityWhen [SequenceGeneratorOption]

GenerationClause ScalarExpr

data IdentityWhen

Constructors

GeneratedAlways

GeneratedByDefault

data SequenceGeneratorOption

Constructors

SGODataType TypeName

SGOStartWith Integer

SGORestart (Maybe Integer)

SGOIncrementBy Integer

SGOMaxValue Integer

SGONoMaxValue

SGOMinValue Integer

SGONoMinValue

SGOCycle

SGONoCycle

data ColConstraintDef

Constructors

ColConstraintDef (Maybe [Name]) ColConstraint

data ColConstraint

Constructors

ColNotNullConstraint

ColUniqueConstraint

ColPrimaryKeyConstraint

ColReferencesConstraint [Name] (Maybe Name) ReferenceMatch ReferentialAction ReferentialAction

ColCheckConstraint ScalarExpr

data TableConstraint

Constructors

TableUniqueConstraint [Name]

TablePrimaryKeyConstraint [Name]

TableReferencesConstraint [Name] [Name] (Maybe [Name]) ReferenceMatch ReferentialAction ReferentialAction

TableCheckConstraint ScalarExpr

data ReferenceMatch

Constructors

DefaultReferenceMatch

MatchFull

MatchPartial

MatchSimple

data ReferentialAction

Constructors

DefaultReferentialAction

RefCascade

RefSetNull

RefSetDefault

RefRestrict

RefNoAction

data AlterTableAction

Constructors

AddColumnDef ColumnDef

AlterColumnSetDefault Name ScalarExpr

AlterColumnDropDefault Name

AlterColumnSetNotNull Name

AlterColumnDropNotNull Name

AlterColumnSetDataType Name TypeName

DropColumn Name DropBehaviour

AddTableConstraintDef (Maybe [Name]) TableConstraint

DropTableConstraintDef [Name] DropBehaviour

data CheckOption

Constructors

DefaultCheckOption

CascadedCheckOption

LocalCheckOption

data AlterDomainAction

Constructors

ADSetDefault ScalarExpr

ADDropDefault

ADAddConstraint (Maybe [Name]) ScalarExpr

ADDropConstraint [Name]

data AdminOption

Constructors

WithAdminOption

WithoutAdminOption

data GrantOption

Constructors

WithGrantOption

WithoutGrantOption

data PrivilegeObject

Constructors

PrivTable [Name]

PrivDomain [Name]

PrivType [Name]

PrivSequence [Name]

PrivFunction [Name]

data PrivilegeAction

Constructors

PrivAll

PrivSelect [Name]

PrivDelete

PrivInsert [Name]

PrivUpdate [Name]

PrivReferences [Name]

PrivUsage

PrivTrigger

PrivExecute

data AdminOptionFor

Constructors

AdminOptionFor

NoAdminOptionFor

data GrantOptionFor

Constructors

GrantOptionFor

NoGrantOptionFor

Comment

data Comment

Comment. Useful when generating SQL code programmatically. The parser doesn't produce these.

Constructors

BlockComment String